Weyl group type C permutations acting on the Young tableau representation of plane trees
This formatted Mathematica 10.1.0 notebook serves as a visualization
aid to “RNA, local moves on plane trees, and transpositions on
tableaux” (http://arxiv.org/abs/1411.3056).

When acting on the Young tableau representation of a set of RNA
secondary structures, or “plane trees,” sharing the same number of
edges, the Weyl group C permutations split these planes trees into two
self-contained orbits, one of symmetric plane trees and the other of
asymmetric plane trees. The visualization aid summarizes the
permutations, Young tableaux, plane trees, and orbits in a layered
graph. The visualization aid also gives a few examples of the effects
of the Weyl group D permutations, which fail to produce valid orbits
and thus are not group actions.